LPS-135347 Move sectionAddedHandler to data-engine-js-components-web
