+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/core/reducers/dragAndDropReducer.es.js
@@ -18,11 +18,11 @@ import {
 	handleResizeLeft,
 	handleResizeRight,
 } from 'dynamic-data-mapping-form-builder/js/components/LayoutProvider/handlers/columnResizedHandler.es';
-import sectionAdded from 'dynamic-data-mapping-form-builder/js/components/LayoutProvider/handlers/sectionAddedHandler.es';
 
 import * as FormSupport from '../../utils/FormSupport.es';
 import {PagesVisitor} from '../../utils/visitors.es';
 import {EVENT_TYPES} from '../actions/eventTypes.es';
+import sectionAdded from '../utils/sectionAddedHandler';
 import {deleteField} from './fieldEditableReducer.es';
 
 const FIELD_TYPES = {
