+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/package.json
@@ -31,7 +31,6 @@
 			"FormRenderer",
 			"LayoutProvider/handlers/columnResizedHandler",
 			"LayoutProvider/handlers/fieldBlurredHandler",
-			"LayoutProvider/handlers/fieldDeletedHandler",
 			"LayoutProvider/handlers/fieldEditedHandler",
 			"LayoutProvider/handlers/fieldSetAddedHandler",
 			"LayoutProvider/handlers/focusedFieldEvaluationEndedHandler",
